move a couple more models to shared package
