+++ packages/client/src/JoinGameModal.tsx
@@ -13,10 +13,10 @@ import {
   Stack,
   FormControl,
   FormLabel,
-  Input
+  Input,
 } from "@chakra-ui/react";
 import React, { useEffect, useState } from "react";
-import { Game } from "./types";
+import { Game } from "shared";
 
 interface Props {
   isOpen: boolean;
@@ -29,7 +29,7 @@ const JoinGameModal: React.FC<Props> = ({
   isOpen,
   findWaitingGames,
   onClose,
-  onJoin
+  onJoin,
 }) => {
   const [games, setGames] = useState<Game[]>([]);
   const [userName, setUserName] = useState("");
@@ -51,7 +51,7 @@ const JoinGameModal: React.FC<Props> = ({
           <FormControl id="userName" isRequired mb="5">
             <FormLabel>Name</FormLabel>
             <Input
-              onChange={e => setUserName(e.target.value)}
+              onChange={(e) => setUserName(e.target.value)}
               placeholder="Peter Davids"
             />
           </FormControl>
