+++ packages/server/src/modules/interfaces.ts
@@ -1,21 +1,5 @@
 import { Card, Deck, Hand } from "typedeck";
-import { GameState } from "shared";
-
-export interface Player {
-  name: string;
-  hand: Hand;
-}
-
-export interface Game {
-  deck: Deck;
-  discardPile: Deck;
-  players: Player[];
-  currentPlayer: number;
-  _id: string;
-  frozen?: boolean;
-  state: GameState;
-  gameName?: string;
-}
+import { Game, GameState } from "shared";
 
 export enum DrawType {
   DECK,
